fix: lob_print now does what it says
It seems like calling lob_print adds a new output each time it gets called
and there were also some issues with what we see in the output and what we see in the actual log.
Now, the logs should match the output, no matter what root loglevel is selected.

The tests have been update to actually catch this issue of multiple outputs of the same print.